refactor(react): move util into utils folder
