+++ packages/react/src/lib/utils/getTextSelectionAncestorPositions.ts
